+++ include/prints.h
@@ -34,12 +34,12 @@ void msgPrintln(bool bleOn, HardwareBLESerial& bleSerial, int64_t msg) {
     if (bleOn) bleSerial.println(msg);
 }
 
-void msgPrint(bool bleOn, HardwareBLESerial& bleSerial, double msg) {
+void msgPrint(bool bleOn, HardwareBLESerial& bleSerial, float msg) {
     Serial.print(msg);
     if (bleOn) bleSerial.print(msg);
 }
 
-void msgPrintln(bool bleOn, HardwareBLESerial& bleSerial, double msg) {
+void msgPrintln(bool bleOn, HardwareBLESerial& bleSerial, float msg) {
     Serial.println(msg);
     if (bleOn) bleSerial.println(msg);
 }
